Move image segment to separate file
